requirements.txt
flask==2.3.2
flask_restful==0.3.10
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
